PP-7588 Move stripe account stubs
