+++ test/cypress/stubs/stripe-account-stubs.js
@@ -1,13 +1,15 @@
 'use strict'
 
-const getStripeAccountSuccess = function (gatewayAccountId, stripeAccountId) {
-  return {
-    name: 'getStripeAccountSuccess',
-    opts: {
-      gateway_account_id: gatewayAccountId,
+const stripeAccountFixtures = require('../../fixtures/stripe-account.fixtures')
+const { stubBuilder } = require('./stub-builder')
+
+function getStripeAccountSuccess (gatewayAccountId, stripeAccountId) {
+  const path = `/v1/api/accounts/${gatewayAccountId}/stripe-account`
+  return stubBuilder('GET', path, 200, {
+    response: stripeAccountFixtures.buildGetStripeAccountResponse({
       stripe_account_id: stripeAccountId
-    }
-  }
+    })
+  })
 }
 
 module.exports = {
